Only try install docker if it doesn't exist

+++ sync.sh
@@ -2,11 +2,13 @@
 THISDIR="$(cd "$(dirname "$0")"; pwd)"
 rsync --archive "$THISDIR/root/" /
 
-dnf -y install dnf-plugins-core
-dnf config-manager \
-    --add-repo \
-    https://download.docker.com/linux/fedora/docker-ce.repo
-dnf -y install docker-ce
+if [[ -f /usr/bin/docker ]]; then
+    dnf -y install dnf-plugins-core
+    dnf config-manager \
+        --add-repo \
+        https://download.docker.com/linux/fedora/docker-ce.repo
+    dnf -y install docker-ce
+fi
 
 systemctl daemon-reload
